Don't break when a project's owner has no repositories on bitbucket.
Normally this should not happen unless our data are out of sync. For
example the owner changes on bitbucket and we still have the old owner
in our database. This handling is only a quick fix because it will leave
the relic project behing and let the user import it again.